Set up

Caculates internal parameters and equilbirum staring values for state variables

set_up(
  LEP = 1,
  delta = 960000,
  HPS = 10000,
  PPS = 2000,
  TPrev = 0.02,
  CPrev = 0.07,
  PCPrev = 0.2,
  AEL = 2,
  ATL = 2,
  ADI = 3,
  LEH = 54,
  phi = 0.8,
  chi = 0.5,
  RR_cysticercosis = 1,
  epsilon = 0.01,
  RR_infection = 1,
  RR_consumption = -0.25
)

Arguments

LEP
Pig life expectancy (years)

delta
Egg production rate (per month)

HPS
Human population size

PPS
Pig population size

TPrev
Taeniasis prevalence

CPrev
Cysticercosis prevalence in humans

PCPrev
Cysticercosis prevalence in pigs

AEL
Average life expectancy of egg in the environment (weeks)

ATL
Average life expectancy of T. solium adult worm in human (years)

ADI
Average duration of cycticercosis infection in humans (years)

LEH
Average life expectancy of a human (years)

phi
Proportion of infected pigs with low-intensity cyst burden

chi
Average number of pork meals (200g portion) per person per month

RR_cysticercosis
Risk multiplier for Cysticercosis if human has Taeniasis

epsilon
Pig rate of loss of naturally aquired immunity

RR_infection
Increased RR of infection given consumption of high cyct burden pork, compared with low cyst burden

RR_consumption
Decreased RR of consumption of high cyst burden pork, compared with low cyst burden pork

Value

Two lists of parameters and state variable values